requirements.txt
fastapi
google-generativea
google-genai
pydantic
python-dotenv
tabulate
termcolor
uvicorn